Not to use second device by default
We build the ami and we would not want to second drive to be attached in the
ami for space efficiency. Thus by default if using overlay2 we just use the
root device.

The option is still there to allow us to apply docker role into existing ec2
(run without AMI scenario)